requirements.txt
discord.py==2.4.0
colorama==0.4.6
python-dotenv==1.1.1